+++ app/Http/Controllers/Backend/BatchExamManagement/BatchExamRoutineController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers\Backend\BatchExamManagement;
 use App\Http\Controllers\Controller;
 use App\Models\Backend\BatchExamManagement\BatchExamRoutine;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+use Symfony\Component\HttpFoundation\Response;
 
 class BatchExamRoutineController extends Controller
 {
@@ -15,7 +17,7 @@ class BatchExamRoutineController extends Controller
      */
     public function index()
     {
-
+        abort_if(Gate::denies('manage-batch-exam-routine'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         if (!empty(request()->input('batch_exam_id')))
         {
             return view('backend.batch-exam-management.batch-exam-routines.index', [
@@ -30,6 +32,7 @@ class BatchExamRoutineController extends Controller
      */
     public function create()
     {
+        abort_if(Gate::denies('create-batch-exam-routine'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         //
     }
 
@@ -38,6 +41,7 @@ class BatchExamRoutineController extends Controller
      */
     public function store(Request $request)
     {
+        abort_if(Gate::denies('store-batch-exam-routine'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $request->validate(['date_time' => 'required','day' => 'required']);
         BatchExamRoutine::createOrUpdateCourseRoutines($request);
         return back()->with('success', 'Batch Exam Routine Created Successfully');
@@ -48,6 +52,7 @@ class BatchExamRoutineController extends Controller
      */
     public function show(string $id)
     {
+        abort_if(Gate::denies('show-batch-exam-routine'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         //
     }
 
@@ -56,6 +61,7 @@ class BatchExamRoutineController extends Controller
      */
     public function edit(string $id)
     {
+        abort_if(Gate::denies('edit-batch-exam-routine'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         return view('backend.batch-exam-management.batch-exam-routines.edit', [
             'batchExamRoutine'   => BatchExamRoutine::find($id),
         ]);
@@ -66,6 +72,7 @@ class BatchExamRoutineController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        abort_if(Gate::denies('update-batch-exam-routine'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         BatchExamRoutine::createOrUpdateCourseRoutines($request, $id);
         return back()->with('success', 'Batch Exam Routine Updated Successfully');
     }
@@ -75,6 +82,7 @@ class BatchExamRoutineController extends Controller
      */
     public function destroy(string $id)
     {
+        abort_if(Gate::denies('delete-batch-exam-routine'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         BatchExamRoutine::find($id)->delete();
         return back()->with('success', 'Batch Exam Routine deleted Successfully');
     }
